requirements.txt
Django==1.11.2
pytz==2017.2
dj-database-url==0.4.2
dj-static==0.0.6
django-toolbelt==0.0.1
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.1
static3==0.7.0
whitenoise==3.3.0